+++ src/components/StackedAreaChart/types.ts
@@ -1,0 +1,9 @@
+import {Color} from 'types';
+
+type Data = number | null;
+
+export interface Series {
+  label: string;
+  data: Data[];
+  color: Color;
+}