+++ src/utilities/index.ts
@@ -2,3 +2,4 @@ export {eventPoint} from './event-point';
 export {getTextWidth} from './get-text-width';
 export {clamp} from './clamp';
 export {getColorValue} from './get-color-value';
+export {uniqueId} from './unique-id';