+++ CHANGELOG.md
@@ -2,11 +2,19 @@
 
 ## Unreleased
 
+### Added
+
+- `<StackedAreaChart />`
+
 ### Fixed
 
 - `<LineChart />` prop improvements to rename `x` to `label` and `y` to `rawValue`
 - `<LineChart />` displays label value on tooltip rather than Series name
 - `<LineChart />` returns null when empty data is passed in
+
+### Changed
+
+- `<Sparkline />` `useAnimation` prop is renamed to `isAnimated`
 
 ## [0.0.15] - 2020-09-10